bot/requirements.txt
python-telegram-bot==21.6
pyairtable==2.3.3
python-dotenv==1.0.1
